feat(UI) – Prepare UI 3.0.0
