+++ src/app/tools/page.tsx
@@ -5,7 +5,7 @@ import { Hr } from '@/components/atoms/hr';
 import { Ul } from '@/components/atoms/ul';
 import type { Metadata } from 'next';
 import { Blockquote } from '@/components/atoms/blockquote';
-import { HeroNavigation } from '@/components/molecules/hero-navigation';
+import { HeroNavigation } from '@/components/molecules/hero-navigation/hero-navigation';
 
 export const metadata: Metadata = {
   title: 'Tools | Dominik Rubröder, UX Design Engineer',
@@ -80,7 +80,7 @@ export default function ToolsPage() {
 
       <Hr variant="container" />
 
-      <HeroNavigation />
+      <HeroNavigation hiddenLink="/tools" />
     </>
   );
 }
