+++ src/components/atoms/markup/index.ts
@@ -1,0 +1,3 @@
+export { Markup } from './markup';
+
+export * from './types';
