+++ src/components/atoms/svg/svg-shapes.tsx
@@ -27,8 +27,10 @@ export function SvgShapes({ id, className }: BaseComponentProps) {
           d="M1407.41 455.414C1408.2 454.633 1408.2 453.367 1407.41 452.586L1394.69 439.858C1393.91 439.077 1392.64 439.077 1391.86 439.858C1391.08 440.639 1391.08 441.905 1391.86 442.686L1403.17 454L1391.86 465.314C1391.08 466.095 1391.08 467.361 1391.86 468.142C1392.64 468.923 1393.91 468.923 1394.69 468.142L1407.41 455.414ZM1468 518C1468 552.242 1440.24 580 1406 580V584C1442.45 584 1472 554.451 1472 518H1468ZM1406 580C1371.76 580 1344 552.242 1344 518H1340C1340 554.451 1369.55 584 1406 584V580ZM1344 518C1344 483.758 1371.76 456 1406 456V452C1369.55 452 1340 481.549 1340 518H1344Z"
           style={{ fill: 'var(--color-primary)' }}
         />
-        <path
-          d="M290.586 412.586C289.805 413.367 289.805 414.633 290.586 415.414L303.314 428.142C304.095 428.923 305.361 428.923 306.142 428.142C306.923 427.361 306.923 426.095 306.142 425.314L294.828 414L306.142 402.686C306.923 401.905 306.923 400.639 306.142 399.858C305.361 399.077 304.095 399.077 303.314 399.858L290.586 412.586ZM666 412H292V416H666V412Z"
+        <circle
+          cx="300"
+          cy="425"
+          r="44"
           style={{ fill: 'var(--color-primary)' }}
         />
         <path
