+++ src/components/atoms/toggle/stories/toggle.stories.tsx
@@ -1,0 +1,40 @@
+import type { Meta, StoryObj } from '@storybook/nextjs-vite';
+import { fn } from 'storybook/test';
+
+import { Toggle } from '@/components/atoms/toggle';
+
+const meta = {
+  title: 'Atoms/Toggle',
+  component: Toggle,
+  parameters: {
+    layout: 'centered'
+  },
+  tags: ['autodocs'],
+  args: {
+    onClick: fn(),
+    isEnabled: false,
+    label: 'Toggle Label'
+  }
+} satisfies Meta<typeof Toggle>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Disabled: Story = {
+  args: {
+    isEnabled: false
+  }
+};
+
+export const Enabled: Story = {
+  args: {
+    isEnabled: true
+  }
+};
+
+export const WithoutLabel: Story = {
+  args: {
+    label: undefined,
+    isEnabled: false
+  }
+};
