+++ src/components/molecules/button-group.tsx
@@ -11,6 +11,7 @@ export function ButtonGroup({
   children
 }: {
   stretchButtons?: boolean;
+  stretchButtonsOnMobile?: boolean;
   minimizeSpacing?: boolean;
   className?: string;
   children: ReactNode;
@@ -20,7 +21,7 @@ export function ButtonGroup({
       className={cn(
         'flex flex-wrap items-center gap-2',
         stretchButtons &&
-          '[&>a]:flex-1 sm:[&>a]:flex-none [&>button]:flex-1 sm:[&>button]:flex-none',
+          'flex-col sm:flex-row [&>a]:w-full [&>a]:flex-1 sm:[&>a]:w-auto sm:[&>a]:flex-none [&>button]:w-full [&>button]:flex-1 sm:[&>button]:w-auto sm:[&>button]:flex-none',
         minimizeSpacing && 'gap-0 -space-x-1',
         className
       )}
