+++ src/components/molecules/hero-navigation/index.ts
@@ -1,0 +1,4 @@
+export { HeroNavigation } from './hero-navigation';
+
+export * from './types';
+export * from './data';
