+++ src/components/molecules/hero-navigation/types.ts
@@ -1,0 +1,8 @@
+export type HeroNavigationPathname =
+  | '/introduction'
+  | '/projects'
+  | '/tools'
+  | '/technologies'
+  | '/showcase'
+  | '/playground'
+  | '/resources';
